requirements.txt
grpcio-tools>1.60
scipy
numpy
faiss-cpu
pandas
redis
protobuf